pkgrec.h: initial revision
Let's design a new format for storing ELF symbols.  ELF symbols come
in compressed records, each record corresponding to a package.  These
records are stored in a single file.  Hence we should be able to read
a record from a file and decompress most, but not all, of its data.
All integers and all LZ4 chunks are decompressed when reading a record;
but sym1 frenc blob won't be unpacked to full-length symbols.  That's
because we don't even need to store all of them.  The symbols from sym1
are identified by symhash, and we may only need to store only the "hi"
part of symhash, whereas the "lo" part is implicitly stored in fp47map
(see bad-proc.c).

The records are preceded by the index, which is conceptually a list
of pairs <pkg-basename, record-size>.  When the record-size is zero,
we know that the package provides no symbols (so what we don't have
to process it again only to find out that it provides no symbols).
Otherwise, we can PkgRec_read() the record.