slab.h: redesigned a bit, factored slab.c
I wanted to try and use mmap/mremap instead of malloc/realloc, but after
looking into glibc/malloc, I realized that I probably don't want to deal
with memory tagging.  But anyway, slab.h should be independent of the
underlying allocator.

A useful property of mmap/mremap is that new memory is zeroed out.
Thus the caller may assume that the rest of the buffer is zero-padded.
This may be useful e.g. when hashing a newly added string.