symhash.c: initial revision
It needs more work, and empirical tests from ../zrha-collisions are
still running on gcc135.fsffrance.org.  But we need to start with
something, to construct a working prototype.

At least I checked that all symbols from elfsym.zst hash to different
values.  The program that does this is similar to
https://github.com/svpv/xxh3-collisions/blob/master/hashline.c

I also checked that, under the same seed, when a symbol gets hashed for
the second time, its hash value is the same. :)

$ (zstd -d <~/elfsym.zst |head -333 && zstd -d <~/elfsym.zst |head -33) |
  ./a.out |sort -u |wc -l
333