symhash.c: update construction to the current best
Here's a glimpse into how I design this hash function.  Searching all
possible combinations is impractical, but can you think of a Cartesian
decomposition?  First, we can find the best way to feed the data F0..F3
(such that minimizes the number of collisions), assuming that all
multiplications fail (we force multiplication by zero).  Then, we find
the best way to shuffle the halves (SHUF0..SHUF1), still assuming that
all multiplication fail.  We select the best way to do SHUF0..SHUF1 and
a few good ways to do F0..F3.  Then we re-enable multiplication and
search for the best way to combine the products with the state F4..F5.

To estimate the number of collisions, I use a real-world data set,
ELF symbols in fact, about 34M keys.  Each line below is a distinct
construction that gets tested 4096 times.  The last two columns show
the total number of lines with collisions (the last column) and the
number of "problematic seeds" under which the collisions are found (the
second-to-last one).

The important thing is that the construction that I'm testing is scaled
down by a factor of four: 8x8->16 multiplication, 64-bit state, 64-bit
output.  'Cause you can't easily find any collisions when you go
32x32->64, 256-bit state, 128-bit output. :-)

The test is still running, and I'm not sure the counts and the differences
between them are statistically significant, but there's only one "0 0"
so far, so why not?

-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Xor  4 8
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Add  5 38
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Sub  8 32
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Xor  5 186
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Add  1 8
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Sub  8 26
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Xor  5 12
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Add  4 94
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Sub  5 48
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Xor  5 18
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Add  9 218
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Sub  9 20
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Xor  5 168
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Add  0 0
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Sub  7 26
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Xor  8 66
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Add  4 90
-DF0=Xor -DF1=Xor -DF2=Add -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Sub  5 56
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Xor  9 42
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Add  5 50
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Sub  4 8
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Xor  1 102
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Add  7 264
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Sub  2 18
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Xor  6 30
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Add  5 102
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Add -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Sub -DF5=Sub  7 16
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Xor  6 48
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Add  6 56
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Xor -DF5=Sub  8 64
-DF0=Xor -DF1=Xor -DF2=Sub -DF3=Sub -DSHUF0=3,2,0,1 -DSHUF1=2,3,1,0 -DF4=Add -DF5=Xor  4 16
...