pkgrec.c print.c: initial revision
This is a "first working version".  It shows how the data can be
represented with 16-bit integers and packaged into a single pfor16 blob.
After pfor16 blob, there goes lz4 blob.  We only use strings when they
are absolutely necessary, i.e. for filenames and for sym1; both filenames
and sym1 are preprocessed with frenc-mini, but the lz4 segment is combined.